Split FLAC track streaming out of cue.go into flac.go.
